refactor: custom hook for handling row actions (save, delete, etc)
